requirements.txt
streamlit
pandas
openpyxl
Pillow
streamlit-js-eval